requirements.txt
# requirements.txt

pandas == 1.5.3

streamlit

geopandas

plotly.graph_objects

json